separate `IntDisjointSet` implementation into disjoint_set.jl
